change type of orderSum from decimal to double

+++ Entities/Order.cs
@@ -9,7 +9,7 @@ public partial class Order
 
     public DateTime? OrderDate { get; set; }
 
-    public decimal OrderSum { get; set; }
+    public double OrderSum { get; set; }
 
     public int UserId { get; set; }
